+++ AdapterUnitTests/LineAdapterTests/AUTLineWaterfallInterstitialAdTest.m
@@ -9,6 +9,7 @@
 
 #import "GADMediationAdapterLineConstants.h"
 #import "GADMediationAdapterLineExtras.h"
+#import "GADMediationAdapterLineUtils.h"
 
 static NSString *const kTestSlotID = @"12345";
 
@@ -25,7 +26,8 @@ static NSString *const kTestSlotID = @"12345";
 
 - (void)setUp {
   [super setUp];
-
+  GADMediationAdapterLineUnregisterFiveAd();
+  
   _adapter = [[GADMediationAdapterLine alloc] init];
 
   _interstitialMock = OCMClassMock([FADInterstitial class]);